server/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile
#
aioredis==1.3.1
    # via channels-redis
asgiref==3.4.1
    # via
    #   channels
    #   channels-redis
    #   daphne
    #   django
    #   djoser
async-timeout==3.0.1
    # via aioredis
attrs==21.2.0
    # via
    #   automat
    #   service-identity
    #   twisted
autobahn==21.3.1
    # via daphne
automat==20.2.0
    # via twisted
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
channels==3.0.4
    # via
    #   -r requirements.in
    #   channels-redis
channels-redis==3.3.1
    # via -r requirements.in
charset-normalizer==2.0.7
    # via requests
constantly==15.1.0
    # via twisted
coreapi==2.3.3
    # via djoser
coreschema==0.0.4
    # via coreapi
cryptography==35.0.0
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
    #   social-auth-core
daphne==3.0.2
    # via channels
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==3.2.8
    # via
    #   -r requirements.in
    #   channels
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cors-headers==3.10.0
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   djangorestframework-gis
    #   djangorestframework-simplejwt
djangorestframework-gis==0.17
    # via -r requirements.in
djangorestframework-simplejwt==4.8.0
    # via djoser
djoser==2.1.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
hiredis==2.0.0
    # via aioredis
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.3
    # via
    #   hyperlink
    #   requests
    #   twisted
importlib-metadata==1.7.0
    # via djoser
incremental==21.3.0
    # via twisted
itypes==1.2.0
    # via coreapi
jinja2==3.0.2
    # via coreschema
markupsafe==2.0.1
    # via jinja2
msgpack==1.0.2
    # via channels-redis
oauthlib==3.1.1
    # via
    #   requests-oauthlib
    #   social-auth-core
psycopg2-binary==2.9.1
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.2.8
    # via service-identity
pycparser==2.20
    # via cffi
pyjwt==2.3.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pyopenssl==21.0.0
    # via twisted
python3-openid==3.2.0
    # via social-auth-core
pytz==2021.3
    # via django
requests==2.26.0
    # via
    #   coreapi
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.0
    # via social-auth-core
service-identity==21.1.0
    # via twisted
six==1.16.0
    # via
    #   automat
    #   pyopenssl
    #   service-identity
    #   social-auth-app-django
social-auth-app-django==4.0.0
    # via djoser
social-auth-core==4.1.0
    # via social-auth-app-django
sqlparse==0.4.2
    # via django
twisted[tls]==21.7.0
    # via daphne
txaio==21.2.1
    # via autobahn
typing-extensions==3.10.0.2
    # via
    #   asgiref
    #   twisted
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.7
    # via requests
zipp==3.6.0
    # via importlib-metadata
zope.interface==5.4.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
